Fold TypeWriter into JavaGenerator.
Now that this functionality is public API (for build tools to use directly) there's
only really one public-facing type. It does both type mapping (int32 to Integer) and
also generating TypeSpec instances.
